[cmt] Get APD calculation to work on sets of traces where the number of APs varies, at least for the examples we're doing so far.
It now pads with the max upstroke time for the first AP in the first run. This is because we're indexing to find when the upstroke occurs, not indexing the final AP calculation, so zero wasn't a good choice - it meant the interpolation couldn't go backwards from that point!  So what you now end up with is negative DI (and replicated APD) where it has padded.

git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/projects/jonc/cell_model_tests@11261 b5f69653-2bf5-0310-8201-b0dafb8e1405