Update reference data for test
pycml made a mistake here I discovered! It didn't convert the
initial value for voltage, so thought the initial condition was
-0.074 mV instead of -74 mV - hence an AP was triggered immediately.

I put the wrong initial condition in to this code and compared the
resulting AP trace with the one on the Web Lab - they look the same.
So am fairly confident this code is now doing the right thing!